fix: use of `SchemaGenerator` for `ZodGenerator`
