+++ packages/swagger-zod/src/components/Query.tsx
@@ -28,14 +28,14 @@ type FileProps = {
 }
 
 Query.File = function({ mode = 'directory' }: FileProps): ReactNode {
-  const { options } = usePlugin<PluginOptions>()
+  const plugin = usePlugin<PluginOptions>()
 
   const schemas = useSchemas()
   const pluginManager = usePluginManager()
   const oas = useOas()
   const file = useOperationFile()
 
-  const builder = new ZodBuilder(options, { oas, pluginManager })
+  const builder = new ZodBuilder(plugin.options, { oas, plugin, pluginManager })
     .add(schemas.pathParams)
     .add(schemas.queryParams)
     .add(schemas.headerParams)
