+++ packages/swagger-zod/src/zodParser.ts
@@ -1,79 +1,6 @@
-export type ZodMetaMapper = {
-  object: { keyword: 'object'; args: { entries: { [x: string]: ZodMeta[] }; strict?: boolean } }
-  strict: { keyword: 'strict' }
-  readOnly: { keyword: 'readOnly' }
-  url: { keyword: 'url' }
-  uuid: { keyword: 'uuid' }
-  email: { keyword: 'email' }
-  date: { keyword: 'date' }
-  datetime: { keyword: 'datetime' }
-  default: { keyword: 'default'; args?: string | number | boolean }
-  lazy: { keyword: 'lazy' }
-  tuple: { keyword: 'tuple'; args?: ZodMeta[] }
-  array: { keyword: 'array'; args?: ZodMeta[] }
-  enum: { keyword: 'enum'; args?: Array<string | number> }
-  and: { keyword: 'and'; args?: ZodMeta[] }
-  literal: { keyword: 'literal'; args: string | number }
-  union: { keyword: 'union'; args?: ZodMeta[] }
-  ref: { keyword: 'ref'; args?: { name: string } }
-  catchall: { keyword: 'catchall'; args?: ZodMeta[] }
-  optional: { keyword: 'optional' }
-  matches: { keyword: 'matches'; args?: string }
-  max: { keyword: 'max'; args?: number }
-  min: { keyword: 'min'; args?: number }
-  describe: { keyword: 'describe'; args?: string }
-  boolean: { keyword: 'boolean' }
-  string: { keyword: 'string' }
-  integer: { keyword: 'integer' }
-  number: { keyword: 'number' }
-  undefined: { keyword: 'undefined' }
-  nullish: { keyword: 'nullish' }
-  nullable: { keyword: 'nullable' }
-  null: { keyword: 'null' }
-  any: { keyword: 'any' }
-  unknown: { keyword: 'unknown' }
-}
+import { isKeyword, schemaKeywords } from '@kubb/swagger'
 
-export const zodKeywords = {
-  any: 'any',
-  unknown: 'unknown',
-  number: 'number',
-  integer: 'integer',
-  object: 'object',
-  lazy: 'lazy',
-  string: 'string',
-  boolean: 'boolean',
-  undefined: 'undefined',
-  nullable: 'nullable',
-  null: 'null',
-  nullish: 'nullish',
-  array: 'array',
-  tuple: 'tuple',
-  enum: 'enum',
-  union: 'union',
-  literal: 'literal',
-  datetime: 'datetime',
-  date: 'date',
-  email: 'email',
-  uuid: 'uuid',
-  url: 'url',
-  strict: 'strict',
-  /* intersection */
-  default: 'default',
-  and: 'and',
-  describe: 'describe',
-  min: 'min',
-  max: 'max',
-  optional: 'optional',
-  catchall: 'catchall',
-  readOnly: 'readOnly',
-
-  // custom ones
-  ref: 'ref',
-  matches: 'matches',
-} satisfies { [K in keyof ZodMetaMapper]: ZodMetaMapper[K]['keyword'] }
-
-export type ZodKeyword = keyof typeof zodKeywords
+import type { Schema, SchemaMapper, SchemaMapperBase } from '@kubb/swagger'
 
 export const zodKeywordMapper = {
   any: 'z.any',
@@ -112,51 +39,42 @@ export const zodKeywordMapper = {
   // custom ones
   ref: 'ref',
   matches: '.regex',
-} satisfies { [K in keyof ZodMetaMapper]: string }
-
-type ZodMetaBase<T> = {
-  keyword: ZodKeyword
-  args: T
-}
-
-export function isKeyword<T extends ZodMeta, K extends keyof ZodMetaMapper>(meta: T, keyword: K): meta is Extract<T, ZodMetaMapper[K]> {
-  return meta.keyword === keyword
-}
-
-export type ZodMeta =
-  | { keyword: string }
-  | ZodMetaMapper[keyof ZodMetaMapper]
+  firstName: undefined,
+  lastName: undefined,
+  password: undefined,
+  phone: undefined,
+} satisfies { [K in keyof SchemaMapper]: string | undefined }
 
 /**
  * @link based on https://github.com/cellular/oazapfts/blob/7ba226ebb15374e8483cc53e7532f1663179a22c/src/codegen/generate.ts#L398
  */
 
-function zodKeywordSorter(a: ZodMeta, b: ZodMeta): 1 | -1 | 0 {
-  if (b.keyword === zodKeywords.null) {
+function zodKeywordSorter(a: Schema, b: Schema): 1 | -1 | 0 {
+  if (b.keyword === schemaKeywords.null) {
     return -1
   }
 
   return 0
 }
 
-export function parseZodMeta(item: ZodMeta = {} as ZodMeta, mapper: Record<ZodKeyword, string> = zodKeywordMapper): string {
+export function parseZodMeta(item: Schema = {} as Schema, mapper: typeof zodKeywordMapper = zodKeywordMapper): string {
   const value = mapper[item.keyword as keyof typeof mapper]
 
-  if (isKeyword(item, zodKeywords.tuple)) {
+  if (isKeyword(item, schemaKeywords.tuple)) {
     return `${value}(${Array.isArray(item.args) ? `[${item.args.map((tupleItem) => parseZodMeta(tupleItem, mapper)).join(',')}]` : parseZodMeta(item.args)})`
   }
 
-  if (isKeyword(item, zodKeywords.enum)) {
+  if (isKeyword(item, schemaKeywords.enum)) {
     return `${value}(${Array.isArray(item.args) ? `[${item.args.join(',')}]` : parseZodMeta(item.args)})`
   }
 
-  if (isKeyword(item, zodKeywords.array)) {
+  if (isKeyword(item, schemaKeywords.array)) {
     return `${value}(${Array.isArray(item.args) ? `${item.args.map((arrayItem) => parseZodMeta(arrayItem, mapper)).join('')}` : parseZodMeta(item.args)})`
   }
-  if (isKeyword(item, zodKeywords.union)) {
+  if (isKeyword(item, schemaKeywords.union)) {
     // zod union type needs at least 2 items
     if (Array.isArray(item.args) && item.args.length === 1) {
-      return parseZodMeta(item.args[0] as ZodMeta)
+      return parseZodMeta(item.args[0] as Schema)
     }
     if (Array.isArray(item.args) && !item.args.length) {
       return ''
@@ -165,24 +83,24 @@ export function parseZodMeta(item: ZodMeta = {} as ZodMeta, mapper: Record<ZodKe
     return `${Array.isArray(item.args) ? `${value}([${item.args.map((unionItem) => parseZodMeta(unionItem, mapper)).join(',')}])` : parseZodMeta(item.args)}`
   }
 
-  if (isKeyword(item, zodKeywords.catchall)) {
+  if (isKeyword(item, schemaKeywords.catchall)) {
     return `${value}(${Array.isArray(item.args) ? `${item.args.map((catchAllItem) => parseZodMeta(catchAllItem, mapper)).join('')}` : parseZodMeta(item.args)})`
   }
 
-  if (isKeyword(item, zodKeywords.and)) {
+  if (isKeyword(item, schemaKeywords.and)) {
     return `${
       item.args
-        ?.filter((item: ZodMeta) => {
-          return ![zodKeywords.optional, zodKeywords.describe].includes(item.keyword as typeof zodKeywords.optional | typeof zodKeywords.describe)
+        ?.filter((item: Schema) => {
+          return ![schemaKeywords.optional, schemaKeywords.describe].includes(item.keyword as typeof schemaKeywords.describe)
         })
-        .map((item: ZodMeta) => parseZodMeta(item, mapper))
+        .map((item: Schema) => parseZodMeta(item, mapper))
         .filter(Boolean)
         .map((item, index) => (index === 0 ? item : `${value}(${item})`))
         .join('')
     }`
   }
 
-  if (isKeyword(item, zodKeywords.object)) {
+  if (isKeyword(item, schemaKeywords.object)) {
     const argsObject = Object.entries(item.args?.entries || '{}')
       .filter((item) => {
         const schema = item[1]
@@ -208,12 +126,12 @@ export function parseZodMeta(item: ZodMeta = {} as ZodMeta, mapper: Record<ZodKe
   }
 
   // custom type
-  if (isKeyword(item, zodKeywords.ref)) {
+  if (isKeyword(item, schemaKeywords.ref)) {
     return `${mapper.lazy}(() => ${item.args?.name})`
   }
 
   if (item.keyword in mapper && 'args' in item) {
-    return `${value}(${(item as ZodMetaBase<unknown>).args as string})`
+    return `${value}(${(item as SchemaMapperBase<unknown>).args as string})`
   }
 
   if (item.keyword in mapper) {
@@ -224,8 +142,8 @@ export function parseZodMeta(item: ZodMeta = {} as ZodMeta, mapper: Record<ZodKe
 }
 
 export function zodParser(
-  items: ZodMeta[],
-  options: { required?: boolean; keysToOmit?: string[]; mapper?: Record<ZodKeyword, string>; name: string; typeName?: string },
+  items: Schema[],
+  options: { required?: boolean; keysToOmit?: string[]; mapper?: typeof zodKeywordMapper; name: string; typeName?: string },
 ): string {
   if (!items.length) {
     return `export const ${options.name} = '';`
