+++ packages/swagger/src/index.ts
@@ -5,6 +5,10 @@ export { OasManager } from './OasManager.ts'
 export type { GetOperationGeneratorOptions, OperationMethodResult } from './OperationGenerator.ts'
 export { OperationGenerator } from './OperationGenerator.ts'
 export { definePlugin, pluginKey, pluginName } from './plugin.ts'
+export type { SchemaGeneratorBuildOptions, SchemaGeneratorOptions } from './SchemaGenerator.ts'
+export { SchemaGenerator } from './SchemaGenerator.ts'
+export type { Schema, SchemaKeyword, SchemaMapper, SchemaMapperBase } from './SchemaMapper.ts'
+export { isKeyword, schemaKeywords } from './SchemaMapper.ts'
 export type * from './types.ts'
 
 /**
